RF: honour configured extractors
With this patch applied, metalad will only
use the extractors defined in datalad.metadata.nativetype,
if any are defined there.

If no extractor is defined in datalad.metadata.nativetype,
it will fall back to metalad_core- and metalad_annex-
extractors.

This allows to prevent full-fledged metadata extraction if
that is not desired.