RF: re-model file and dataset level extraction
This is the first step to separating file- and
dataset level extractors.

The file level extraction is now done by specific
file-level extractors. They receive only a single
file as input. To generate metadata for all files
of a dataset, datalad has to be called on all
files of the dataset.

The dataset level extraction is done by specific
dataset level extractors. They get a complete
dataset as input and are free to generate results
for all elements in the dataset, but all
results are associated with the dataset level
metadata. I.e. if you want to associate results
with individual files, you have to use a file-level
extractor.

Therefore the metalad_core extractors are now
devided into metalad_core_dataset and
metalad_core_file.